Refactored code to conform to Redux file structure standards
