+++ package.json
@@ -24,6 +24,9 @@
     "react-native": "^0.49.5",
     "react-native-elements": "^0.18.2",
     "react-navigation": "^1.0.0-beta.19",
+    "react-redux": "^5.0.6",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.2.0",
     "util": "^0.10.3"
   }
 }
